fix: chart key for bars names
