+++ front/src/features/stats/Chart.js
@@ -1,25 +1,13 @@
 import React from 'react';
 import { useTheme } from '@material-ui/core/styles';
-import { XAxis, YAxis, Label, ResponsiveContainer,BarChart, Bar, CartesianGrid, Tooltip, Legend, } from 'recharts';
+import { XAxis, YAxis, Label, ResponsiveContainer, BarChart, Bar, CartesianGrid, Tooltip, Legend, } from 'recharts';
 import Title from './Title';
 import PropTypes from 'prop-types'
+import { stringToColour } from './stringToColour';
 
-function  stringToColour(str='') {
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-      hash = str.charCodeAt(i) + ((hash << 5) - hash);
-  }
-  let colour = '#';
-  for (let i = 0; i < 3; i++) {
-      let value = (hash >> (i * 8)) & 0xFF;
-      colour += ('00' + value.toString(16)).substr(-2);
-  }
-  return colour;
-}
-
-export default function Chart({sessions, title, names}) {
+export default function Chart({ sessions, title, names }) {
   const theme = useTheme();
-  const Bars = names.map (x=>  <Bar dataKey={x} name={x} stackId='a' fill={stringToColour(x)} />)
+  const Bars = names.map(x => <Bar key={x} dataKey={x} name={x} stackId='a' fill={stringToColour(x)} />)
   return (
     <React.Fragment>
       <Title>{title}</Title>
@@ -43,8 +31,8 @@ export default function Chart({sessions, title, names}) {
             </Label>
           </YAxis>
           <CartesianGrid strokeDasharray="3 3" />
-          <Legend/>
-          <Tooltip/>
+          <Legend />
+          <Tooltip />
           {Bars}
         </BarChart>
       </ResponsiveContainer>
@@ -54,11 +42,10 @@ export default function Chart({sessions, title, names}) {
 
 Chart.propTypes = {
   title: PropTypes.string,
+  names: PropTypes.arrayOf(PropTypes.string).isRequired,
   sessions: PropTypes.arrayOf(PropTypes.shape({
-    duration: PropTypes.number.isRequired,
     startDate: PropTypes.string.isRequired,
-    name: PropTypes.string,
-  }),),
+  })),
 };
 Chart.defaultProps = {
   sessions: []
